# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /source/validate:
    get:
      operationId: getSourceValidate
      tags: [source-validate]
      summary: Validate a Verified Introduction source id (read-only)
      description: >-
        Read-only validation of a client-supplied bytes32 source id against
        SourceRegistryV1 (sourceExists + isActive). The source id is NEVER
        echoed back, and no wallets, addresses, terms, or economic fields from
        the on-chain source configuration are ever returned — only existence
        and active-state booleans. Fails closed: on wrong chain, missing code,
        or decode failure the affected value is null with a failureReason.
      parameters:
        - name: sourceId
          in: query
          required: true
          description: bytes32 hex source id (0x + 64 hex). Never echoed back.
          schema:
            type: string
            maxLength: 66
      responses:
        "200":
          description: Read-only source validation result (source id never echoed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceValidateResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThrottledError"
...
components:
  schemas:
    ThrottledError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SourceValidateResponse:
      type: object
      properties:
        mode:
          type: string
          enum:
            - READ_ONLY_SOURCE_VALIDATE
        asOf:
          type: string
        formatValid:
          type: boolean
        chainVerified:
          type: boolean
        registryCodePresent:
          anyOf:
            - type: boolean
            - type: "null"
        exists:
          anyOf:
            - type: boolean
            - type: "null"
        active:
          anyOf:
            - type: boolean
            - type: "null"
        failureReason:
          anyOf:
            - type: string
            - type: "null"
      required:
        - mode
        - asOf
        - formatValid
        - chainVerified
        - registryCodePresent
        - exists
        - active
        - failureReason
...